add maintenance status note to readme (#47)
* add maintenance status note to readme

@alex-saunders We're adding maintenance status to all our OSS. I think this project should be listed as "Active", but if you have no new featured planned for it, we could list it as "Stable" instead. If you think that's more appropriate, please replace the note I've added with:
```
Maintenance Status: Stable

Formidable is not planning to develop any new features for this project. We are still responding to bug reports and security concerns. We are still welcoming PRs for this project, but PRs that include new features should be small and easy to integrate and should not include breaking changes.
```

Let me know if you have any questions

* Update README.md